@eitsupi my apologies, I didn't realize my local test was pulling the
upstream image. The tex issue is specific to the changes in #743 --
specifically, the official ubuntu repo texlive binaries are installed
and purged in the BUILDDEPS list. By not purging builddeps, we had
conflicting versions of texlive.

This PR modifies the `install_texlive.sh` to not do the manual install
of tlmgr from CPAN if the texlive has already be installed by apt-get
(i.e. `/usr/bin/latex` already exists). I believe it should resolve the
issue.

+++ scripts/install_verse.sh
@@ -74,9 +74,12 @@ apt-get remove -y systemd
 apt-get -y autoremove
 
 ## Add LaTeX, rticles and bookdown support
-wget "https://travis-bin.yihui.name/texlive-local.deb"
-dpkg -i texlive-local.deb
-rm texlive-local.deb
+## tinytex recommends a dummy texlive if using tlmgr manually
+if [[ ! -x "/usr/bin/latex" ]]; then
+    wget "https://travis-bin.yihui.name/texlive-local.deb"
+    dpkg -i texlive-local.deb
+    rm texlive-local.deb
+fi
 
 ## Install texlive
 /rocker_scripts/install_texlive.sh
